all-requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
attrs==22.2.0
black==23.1.0
cachetools==5.5.1
-e catanatron_core
-e catanatron_gym
-e catanatron_server
-e catanatron_experimental
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coverage==6.5.0
coveralls==3.3.1
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
eval_type_backport==0.2.2
Farama-Notifications==0.0.4
filelock==3.17.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==20.1.0
gymnasium==0.29.1
gymnasium-notices==0.0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
jax-jumpy==1.0.0
Jinja2==3.1.2
jiter==0.8.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain-anthropic==0.3.7
langchain-core==0.3.37
langchain-google-vertexai==2.0.13
langchain-mistralai==0.2.6
langchain-openai==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.8
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mistralai==1.5.0
msgpack==1.1.0
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
openai==1.63.2
orjson==3.10.15
packaging==24.1
pandas==1.5.3
pathspec==0.11.1
platformdirs==3.1.1
pluggy==1.0.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.5
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.14.0
pyparsing==3.2.1
pytest==7.2.2
pytest-benchmark==4.0.0
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.2
requests-toolbelt==1.0.0
rich==13.3.2
rsa==4.9
shapely==2.0.7
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.7
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.15
uuid==1.30
watchdog==2.3.1
Werkzeug==2.2.3
zstandard==0.23.0
